move PerfLogger
